Introduces the Matching DSL.
The Matching DSL is a new post-processing API that allows users to match a
specimen builder to incoming requests based on a set of criteria. These
criteria can be based on type variance, member names or a combination of both.

The API is built on top of the IRequestSpecification abstraction and its
entry-point is available as an extension of IPostprocessComposer<T>.